Correct the add_game_module import statement
